+++ src/main/java/de/gematik/test/tiger/fhir/validation/staticv/StaticFhirValidation.java
@@ -18,10 +18,9 @@ package de.gematik.test.tiger.fhir.validation.staticv;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import ca.uhn.fhir.validation.ResultSeverityEnum;
 import ca.uhn.fhir.validation.SingleValidationMessage;
 import de.gematik.rbellogger.data.RbelElement;
-import de.gematik.refv.SupportedValidationModule;
-import de.gematik.refv.ValidationModuleFactory;
 import de.gematik.refv.commons.validation.ValidationMessagesFilter;
 import de.gematik.refv.commons.validation.ValidationModule;
 import de.gematik.refv.commons.validation.ValidationOptions;
@@ -31,10 +30,9 @@ import io.cucumber.core.plugin.report.Evidence;
 import io.cucumber.core.plugin.report.Evidence.Type;
 import io.cucumber.core.plugin.report.EvidenceRecorder;
 import io.cucumber.core.plugin.report.EvidenceRecorderFactory;
-import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
-import lombok.SneakyThrows;
+import java.util.stream.Collectors;
+
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -45,48 +43,48 @@ import org.jetbrains.annotations.NotNull;
 public class StaticFhirValidation {
 
   private static final String BODY_RBEL_PATH = "$.body";
-  private final Map<SupportedValidationModule, ValidationModule> validationModuleCache =
-      new EnumMap<>(SupportedValidationModule.class);
 
   private static final RbelMessageValidator rbelValidator = RbelMessageValidator.instance;
 
   private final EvidenceRecorder evidenceRecorder = EvidenceRecorderFactory.getEvidenceRecorder();
 
-  public void tgrCurrentRequestBodyAtIsValidFHIRRessourceOfType(
-      final SupportedValidationModule validationType) {
-    tgrCurrentRequestAtIsValidFHIRRessourceOfType(BODY_RBEL_PATH, validationType);
+  public void tgrCurrentRequestBodyAtIsValidFHIRResourceOfType(
+          final ValidationModule validationModule) {
+    tgrCurrentRequestAtIsValidFHIRResourceOfType(BODY_RBEL_PATH, validationModule);
   }
 
-  public void tgrCurrentRequestBodyAtIsValidFHIRRessourceOfType(
-      SupportedValidationModule validationType, String profileUrl) {
-    tgrCurrentRequestAtIsValidFHIRRessourceOfType(
-        BODY_RBEL_PATH, validationType, createValidationOptionsWithProfile(profileUrl));
+  public void tgrCurrentRequestBodyAtIsValidFHIRResourceOfType(
+          ValidationModule validationModule, String profileUrl) {
+    tgrCurrentRequestAtIsValidFHIRResourceOfType(
+            BODY_RBEL_PATH, validationModule, createValidationOptionsWithProfile(profileUrl));
   }
 
-  public void tgrCurrentRequestAtIsValidFHIRRessourceOfType(
-      final String rbelPath, final SupportedValidationModule validationType) {
+  public void tgrCurrentRequestAtIsValidFHIRResourceOfType(
+          final String rbelPath, final ValidationModule validationModule) {
     ValidationOptions validationOptions = ValidationOptions.getDefaults();
     validationOptions.setValidationMessagesFilter(ValidationMessagesFilter.KEEP_ALL);
-    tgrCurrentRequestAtIsValidFHIRRessourceOfType(rbelPath, validationType, validationOptions);
+    tgrCurrentRequestAtIsValidFHIRResourceOfType(rbelPath, validationModule, validationOptions);
   }
 
-  public void tgrCurrentRequestAtIsValidFHIRRessourceOfType(
-      String rbelPath, SupportedValidationModule validationType, String profileUrl) {
-    tgrCurrentRequestAtIsValidFHIRRessourceOfType(
-        rbelPath, validationType, createValidationOptionsWithProfile(profileUrl));
+  public void tgrCurrentRequestAtIsValidFHIRResourceOfType(
+          String rbelPath, ValidationModule validationModule, String profileUrl) {
+    tgrCurrentRequestAtIsValidFHIRResourceOfType(
+            rbelPath, validationModule, createValidationOptionsWithProfile(profileUrl));
   }
 
-  private void tgrCurrentRequestAtIsValidFHIRRessourceOfType(
-      String rbelPath,
-      SupportedValidationModule validationType,
-      ValidationOptions validationOptions) {
+  private void tgrCurrentRequestAtIsValidFHIRResourceOfType(
+          String rbelPath,
+          ValidationModule validationModule,
+          ValidationOptions validationOptions) {
     final var validationResult =
-        validateRbelElementAt(
-            rbelValidator.findElementInCurrentRequest(rbelPath), validationType, validationOptions);
+            validateRbelElementAt(
+                    rbelValidator.findElementInCurrentRequest(rbelPath), validationModule, validationOptions);
+
+    final String validationMessage = getCombinedValidationMessages(validationResult);
 
     assertThat(validationResult.isValid())
-        .withFailMessage(() -> "request body is not a valid %s resource".formatted(validationType))
-        .isTrue();
+            .withFailMessage("request body is not a valid %s resource. cause: \n%s", validationModule.getId(), validationMessage)
+            .isTrue();
   }
 
   private static ValidationOptions createValidationOptionsWithProfile(String profile) {
@@ -98,84 +96,81 @@ public class StaticFhirValidation {
 
   private void recordEvidencesOf(final ValidationResult validationResult) {
     validationResult.getValidationMessages().stream()
-        .map(StaticFhirValidation::toEvidence)
-        .forEach(evidenceRecorder::recordEvidence);
+            .map(StaticFhirValidation::toEvidence)
+            .forEach(evidenceRecorder::recordEvidence);
   }
 
   @NotNull
   private static Evidence toEvidence(final SingleValidationMessage it) {
     final Type type =
-        switch (it.getSeverity()) {
-          case INFORMATION -> Type.INFO;
-          case WARNING -> Type.WARN;
-          case ERROR -> Type.ERROR;
-          case FATAL -> Type.FATAL;
-        };
+            switch (it.getSeverity()) {
+              case INFORMATION -> Type.INFO;
+              case WARNING -> Type.WARN;
+              case ERROR -> Type.ERROR;
+              case FATAL -> Type.FATAL;
+            };
 
     return new Evidence(type, it.getMessage(), it);
   }
 
-  public void tgrCurrentResponseBodyAtIsValidFHIRRessourceOfType(
-      final SupportedValidationModule validationType) {
-    tgrCurrentResponseAtIsValidFHIRRessourceOfType(BODY_RBEL_PATH, validationType);
+  public void tgrCurrentResponseBodyAtIsValidFHIRResourceOfType(
+          final ValidationModule validationModule) {
+    tgrCurrentResponseAtIsValidFHIRResourceOfType(BODY_RBEL_PATH, validationModule);
   }
 
-  public void tgrCurrentResponseAtIsValidFHIRRessourceOfType(
-      final String rbelPath, final SupportedValidationModule validationType) {
+  public void tgrCurrentResponseAtIsValidFHIRResourceOfType(
+          final String rbelPath, final ValidationModule validationModule) {
     ValidationOptions validationOptions = ValidationOptions.getDefaults();
     validationOptions.setValidationMessagesFilter(ValidationMessagesFilter.KEEP_ALL);
-    tgrCurrentResponseAtIsValidFHIRRessourceOfType(rbelPath, validationType, validationOptions);
+    tgrCurrentResponseAtIsValidFHIRResourceOfType(rbelPath, validationModule, validationOptions);
   }
 
-  public void tgrCurrentResponseBodyAtIsValidFHIRRessourceOfType(
-      SupportedValidationModule validationType, String profileUrl) {
-    tgrCurrentResponseAtIsValidFHIRRessourceOfType(
-        BODY_RBEL_PATH, validationType, createValidationOptionsWithProfile(profileUrl));
+  public void tgrCurrentResponseBodyAtIsValidFHIRResourceOfType(
+          ValidationModule validationModule, String profileUrl) {
+    tgrCurrentResponseAtIsValidFHIRResourceOfType(
+            BODY_RBEL_PATH, validationModule, createValidationOptionsWithProfile(profileUrl));
   }
 
-  public void tgrCurrentResponseAtIsValidFHIRRessourceOfType(
-      String rbelPath, SupportedValidationModule validationType, String profileUrl) {
-    tgrCurrentResponseAtIsValidFHIRRessourceOfType(
-        rbelPath, validationType, createValidationOptionsWithProfile(profileUrl));
+  public void tgrCurrentResponseAtIsValidFHIRResourceOfType(
+          String rbelPath, ValidationModule validationModule, String profileUrl) {
+    tgrCurrentResponseAtIsValidFHIRResourceOfType(
+            rbelPath, validationModule, createValidationOptionsWithProfile(profileUrl));
   }
 
-  private void tgrCurrentResponseAtIsValidFHIRRessourceOfType(
-      String rbelPath,
-      SupportedValidationModule validationType,
-      ValidationOptions validationOptions) {
+  private void tgrCurrentResponseAtIsValidFHIRResourceOfType(
+          String rbelPath,
+          ValidationModule validationModule,
+          ValidationOptions validationOptions) {
     final var validationResult =
-        validateRbelElementAt(
-            rbelValidator.findElementInCurrentResponse(rbelPath),
-            validationType,
-            validationOptions);
+            validateRbelElementAt(
+                    rbelValidator.findElementInCurrentResponse(rbelPath),
+                    validationModule,
+                    validationOptions);
+
+    final String validationMessage = getCombinedValidationMessages(validationResult);
 
     assertThat(validationResult.isValid())
-        .withFailMessage(() -> "response body is not a valid %s resource".formatted(validationType))
-        .isTrue();
+            .withFailMessage("response body is not a valid %s resource, cause: \n%s", validationModule.getId(), validationMessage)
+            .isTrue();
   }
 
   @NotNull
   private ValidationResult validateRbelElementAt(
-      RbelElement toValidate,
-      SupportedValidationModule validationType,
-      ValidationOptions validationOptions) {
+          RbelElement toValidate,
+          ValidationModule validationModule,
+          ValidationOptions validationOptions) {
     final var rawContentToValidate = toValidate.getRawStringContent();
 
-    final var validator = getValidationModuleFor(validationType);
-    final var validationResult = validator.validateString(rawContentToValidate, validationOptions);
+    final var validationResult = validationModule.validateString(rawContentToValidate, validationOptions);
 
     recordEvidencesOf(validationResult);
     return validationResult;
   }
 
-  private ValidationModule getValidationModuleFor(final SupportedValidationModule validationType) {
-    return validationModuleCache.computeIfAbsent(
-        validationType, StaticFhirValidation::getNewValidationModule);
-  }
-
-  @SneakyThrows
-  private static ValidationModule getNewValidationModule(
-      final SupportedValidationModule validationType) {
-    return new ValidationModuleFactory().createValidationModule(validationType);
+  private String getCombinedValidationMessages(ValidationResult validationResult) {
+    return validationResult.getValidationMessages().stream()
+            .filter(message -> message.getSeverity().equals(ResultSeverityEnum.ERROR))
+            .map(message -> message.getSeverity() + " - " + message.getMessage())
+            .collect(Collectors.joining("\n"));
   }
 }