+++ src/test/java/de/gematik/test/tiger/glue/fhir/FhirPathValidationGlueTest.java
@@ -18,9 +18,14 @@ package de.gematik.test.tiger.glue.fhir;
 
 import static de.gematik.test.tiger.glue.fhir.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
-
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.atLeastOnce;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
 import de.gematik.test.tiger.fhir.validation.fhirpath.FhirPathValidation;
 import de.gematik.test.tiger.fhir.validation.fhirpath.NetTracer;
 import io.cucumber.core.plugin.report.Evidence;
@@ -67,7 +72,8 @@ class FhirPathValidationGlueTest {
   @DisplayName("Empty request body should serve an appropriate error")
   void emptyRequestBodyShouldServeAnAppropriateError() {
     // Arrange
-    when(netTracer.getCurrentRequestsRawStringByRbelPath("$.body")).thenReturn(Optional.empty());
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY)).thenReturn(
+        Optional.empty());
     when(netTracer.getCurrentRequestsRawStringByRbelPath("$.headers.content-type"))
         .thenReturn(Optional.of("application/fhir+json"));
 
@@ -90,7 +96,8 @@ class FhirPathValidationGlueTest {
   @DisplayName("Empty response body should serve an appropriate error")
   void emptyResponseBodyShouldServeAnAppropriateError() {
     // Arrange
-    when(netTracer.getCurrentResponseRawStringByRbelPath("$.body")).thenReturn(Optional.empty());
+    when(netTracer.getCurrentResponseRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY)).thenReturn(
+        Optional.empty());
     when(netTracer.getCurrentResponseRawStringByRbelPath("$.headers.content-type"))
         .thenReturn(Optional.of("application/fhir+json"));
 
@@ -114,7 +121,7 @@ class FhirPathValidationGlueTest {
   @SneakyThrows
   void fhirPathValidationJsonRequest_MissingContentHeaderInResponseShouldServeAnAppropriateError() {
     // Arrange
-    when(netTracer.getCurrentRequestsRawStringByRbelPath("$.body"))
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
         .thenReturn(Optional.of(getFhirRessourceAsJson()));
     final String fhirPath = getTruthyResultFhirPath();
 
@@ -135,9 +142,9 @@ class FhirPathValidationGlueTest {
   void
       fhirPathValidationJsonResponse_MissingContentHeaderInResponseShouldServeAnAppropriateError() {
     // Arrange
-    when(netTracer.getCurrentResponseRawStringByRbelPath("$.body"))
+    when(netTracer.getCurrentResponseRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
         .thenReturn(Optional.of(getFhirRessourceAsJson()));
-    when(netTracer.getCurrentResponseRawStringByRbelPath("$.header.Content-Type"))
+    when(netTracer.getCurrentResponseRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
         .thenReturn(Optional.empty());
     final String fhirPath = getTruthyResultFhirPath();
 
@@ -333,7 +340,7 @@ class FhirPathValidationGlueTest {
     assertThatThrownBy(
             () ->
                 underTest.tgrCurrentResponseEvaluatesTheFhirPath(
-                    "$.body", fhirPath, "Achtung diese Fehlermeldung wird gezeigt"))
+                    FhirPathValidation.RBEL_SELECTOR_FOR_BODY, fhirPath, "Achtung diese Fehlermeldung wird gezeigt"))
         .isInstanceOf(AssertionError.class)
         .hasMessageContaining("Achtung diese Fehlermeldung wird gezeigt");
   }
@@ -371,7 +378,7 @@ class FhirPathValidationGlueTest {
     assertThatThrownBy(
             () ->
                 underTest.tgrCurrentRequestEvaluatesTheFhirPath(
-                    "$.body", fhirPath, "Achtung diese Fehlermeldung wird gezeigt"))
+                    FhirPathValidation.RBEL_SELECTOR_FOR_BODY, fhirPath, "Achtung diese Fehlermeldung wird gezeigt"))
         .isInstanceOf(AssertionError.class)
         .hasMessageContaining("Achtung diese Fehlermeldung wird gezeigt");
   }
@@ -655,6 +662,159 @@ class FhirPathValidationGlueTest {
                                     it -> assertTrue(((BooleanType) it).booleanValue()))));
   }
 
+  @Test
+  @DisplayName("at JSON Response should evaluate valid fhir path expressions and store it to a variable - one boolean result")
+  @SneakyThrows
+  void fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString_ShouldEvaluateOneValidFhirPathExpressions() {
+    // Arrange
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
+        .thenReturn(Optional.of(getFhirRessourceAsJson()));
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
+        .thenReturn(Optional.of("application/json"));
+    final String fhirPath = getTruthyResultFhirPath();
+
+    String variable = "keyVar";
+    // Act
+    underTest.fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString(fhirPath,
+        variable);
+    // Assert
+    org.assertj.core.api.Assertions.assertThat(TigerGlobalConfiguration.readString(variable)).isEqualTo("true");
+  }
+
+  @Test
+  @DisplayName("at JSON Response should evaluate valid fhir path expressions and store it to a variable - one int result")
+  @SneakyThrows
+  void fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString_ShouldEvaluateOneIntValidFhirPathExpressions() {
+    // Arrange
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
+        .thenReturn(Optional.of(getFhirRessourceAsJson()));
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
+        .thenReturn(Optional.of("application/json"));
+    final String fhirPath = getCountResultFhirPath();
+
+    String variable = "keyVar";
+    // Act
+    underTest.fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString(fhirPath,
+        variable);
+    // Assert
+    org.assertj.core.api.Assertions.assertThat(TigerGlobalConfiguration.readString(variable)).isEqualTo("1");
+  }
+
+  @Test
+  @DisplayName("at JSON Response should evaluate valid fhir path expressions and store it to a variable - one dateAndTime result")
+  @SneakyThrows
+  void fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString_ShouldEvaluateOneDateValidFhirPathExpressions() {
+    // Arrange
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
+        .thenReturn(Optional.of(getFhirRessourceAsJson()));
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
+        .thenReturn(Optional.of("application/json"));
+    final String fhirPath = getDateResultFhirPath();
+
+    String variable = "keyVar";
+    // Act
+    underTest.fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString(fhirPath,
+        variable);
+    // Assert
+    org.assertj.core.api.Assertions.assertThat(TigerGlobalConfiguration.readString(variable))
+        .isEqualTo("2022-05-20T08:00:00Z");
+  }
+
+  @Test
+  @DisplayName("at JSON Response should evaluate valid fhir path expressions and store it to a variable - two results")
+  @SneakyThrows
+  void fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString_ShouldEvaluateTwoResultsValidFhirPathExpressions() {
+    // Arrange
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
+        .thenReturn(Optional.of(getFhirRessourceAsJson()));
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
+        .thenReturn(Optional.of("application/json"));
+    final String fhirPath = getMoreResultFhirPath();
+
+    String variable = "keyVar";
+    // Act
+    underTest.fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString(fhirPath,
+        variable);
+    // Assert
+    org.assertj.core.api.Assertions.assertThat(TigerGlobalConfiguration.readString(variable)).isEqualTo("Practitioner");
+  }
+
+  @Test
+  @DisplayName("at JSON Response should evaluate valid fhir path expressions and store it to a variable - complex result")
+  @SneakyThrows
+  void fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString_ShouldEvaluateComplexFhirPathExpressions() {
+    // Arrange
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
+        .thenReturn(Optional.of(getFhirRessourceAsJson()));
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
+        .thenReturn(Optional.of("application/json"));
+    final String fhirPath = getComplexResultFhirPath();
+
+    String variable = "keyVar";
+    // Act
+    final var thrownException =
+        assertThrows(
+            AssertionError.class,
+            () -> underTest.fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString(
+                fhirPath, variable));
+
+    // Assert
+    final String complexFhirPathErrorMessage =
+        "result is not a primitive type for FHIRPath: " + getComplexResultFhirPath();
+    org.assertj.core.api.Assertions.assertThat(thrownException)
+        .hasMessageContainingAll(complexFhirPathErrorMessage);
+  }
+
+  @Test
+  @DisplayName("at JSON Response should evaluate valid fhir path expressions and store it to a variable - invalid FHIRPath")
+  @SneakyThrows
+  void fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString_ShouldFailWithInvalidFhirPathExpressions() {
+    // Arrange
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
+        .thenReturn(Optional.of(getFhirRessourceAsJson()));
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
+        .thenReturn(Optional.of("application/json"));
+    final String fhirPath = getInvalidFhirPath();
+
+    String variable = "keyVar";
+    // Act
+    final var thrownException =
+        assertThrows(
+            AssertionError.class,
+            () -> underTest.fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString(
+                fhirPath, variable));
+
+    // Assert
+    final String invalidFhirPathErrorMessage = "invalid FHIRPath: " + getInvalidFhirPath();
+    org.assertj.core.api.Assertions.assertThat(thrownException)
+        .hasMessageContainingAll(invalidFhirPathErrorMessage);
+  }
+
+  @Test
+  @DisplayName("at JSON Response should evaluate valid fhir path expressions and store it to a variable - empty result")
+  @SneakyThrows
+  void fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString_ShouldFailWithEmptyFhirPathExpressions() {
+    // Arrange
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
+        .thenReturn(Optional.of(getFhirRessourceAsJson()));
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
+        .thenReturn(Optional.of("application/json"));
+    final String fhirPath = getEmptyResultFhirPath();
+
+    String variable = "";
+    // Act
+    final var thrownException =
+        assertThrows(
+            AssertionError.class,
+            () -> underTest.fhirEvaluateFHIRPathTigerResolvedStringOnCurrentResponseBodyAndStoreResultInVariableString(
+                fhirPath, variable));
+
+    // Assert
+    final String emptyFhirPathErrorMessage = "result is null or empty for FHIRPath: " + getEmptyResultFhirPath();
+    org.assertj.core.api.Assertions.assertThat(thrownException)
+        .hasMessageContainingAll(emptyFhirPathErrorMessage);
+  }
+
   @NotNull
   private static String getInvalidFhirPath() {
     return "%&/()=?";
@@ -676,29 +836,49 @@ class FhirPathValidationGlueTest {
   }
 
   private void requestServesValidXmlRessources(String fhirResource) {
-    when(netTracer.getCurrentRequestsRawStringByRbelPath("$.body"))
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
         .thenReturn(Optional.of(fhirResource));
-    when(netTracer.getCurrentRequestsRawStringByRbelPath("$.header.Content-Type"))
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
         .thenReturn(Optional.of("application/xml"));
   }
 
   private void requestServesValidJsonRessources(String fhirResource) {
-    when(netTracer.getCurrentRequestsRawStringByRbelPath("$.body"))
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
         .thenReturn(Optional.of(fhirResource));
-    when(netTracer.getCurrentRequestsRawStringByRbelPath("$.header.Content-Type"))
+    when(netTracer.getCurrentRequestsRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
         .thenReturn(Optional.of("application/json"));
   }
 
   private void responseServesValidXmlRessources(String fhirResource) {
-    when(netTracer.getCurrentResponseRawStringByRbelPath("$.body"))
+    when(netTracer.getCurrentResponseRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_BODY))
         .thenReturn(Optional.of(fhirResource));
-    when(netTracer.getCurrentResponseRawStringByRbelPath("$.header.Content-Type"))
+    when(netTracer.getCurrentResponseRawStringByRbelPath(FhirPathValidation.RBEL_SELECTOR_FOR_CONTENT_TYPE_HEADER))
         .thenReturn(Optional.of("application/xml"));
   }
 
   @NotNull
   private static String getTruthyResultFhirPath() {
     return "Bundle.entry.resource.author.type.where(value = \"Device\").exists()";
+  }
+
+  @NotNull
+  private static String getCountResultFhirPath() {
+    return "Bundle.entry.resource.author.type.where(value = \"Device\").count()";
+  }
+
+  @NotNull
+  private static String getMoreResultFhirPath() {
+    return "Bundle.entry.resource.author.type.where(value != \"Practitioner\" or value != \"Device\")";
+  }
+
+  @NotNull
+  private static String getDateResultFhirPath() {
+    return "Bundle.entry.resource.date";
+  }
+
+  @NotNull
+  private static String getComplexResultFhirPath() {
+    return "Bundle.entry.resource.address";
   }
 
   @NotNull